removed the type cast in sizze

+++ 0x1E-search_algorithms/0-linear.c
@@ -14,7 +14,7 @@ int linear_search(int *array, size_t size, int value)
 	if (!array || size == 0)
 		return (-1);
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		printf("Value checked array [%lu] = [%d]\n", i, array[i]);
 		if (array[i] == value)
